+++ lib/additionals/patches/issue_patch.rb
@@ -34,9 +34,12 @@ module Additionals
         def autowatch_involved
           return unless Additionals.settings[:issue_autowatch_involved].to_i == 1
 
-          add_autowatcher(author) if new_record? || author_id != author_id_was
           add_autowatcher(User.current)
-          add_autowatcher(assigned_to) unless assigned_to.nil? || assigned_to.id == User.current.id
+          add_autowatcher(author) if new_record? || author_id != author_id_was
+          unless assigned_to_id.nil? || assigned_to_id == User.current.id
+            #raise "assigned_to_id: #{assigned_to_id.inspect} - #{assigned_to_id_was.inspect}"
+            add_autowatcher(assigned_to) if new_record? || assigned_to_id != assigned_to_id_was
+          end
 
           true
         end