+++ test/test_helper.rb
@@ -1,10 +1,8 @@
 require 'simplecov'
 require 'simplecov-rcov'
 
-SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
-  SimpleCov::Formatter::HTMLFormatter,
-  SimpleCov::Formatter::RcovFormatter
-]
+SimpleCov.formatters = SimpleCov::Formatter::MultiFormatter[SimpleCov::Formatter::HTMLFormatter,
+                                                            SimpleCov::Formatter::RcovFormatter]
 
 SimpleCov.start :rails do
   add_filter 'init.rb'